+++ test/benchmark.js
@@ -7,102 +7,60 @@
  *
  * You can also specify a specific version to compare with the current version,
  * for example:
- *   `npm run compare-performance -- 2.14.0`
+ *   `npm run compare-performance -- -v 2.14.0`
+ *
+ * To see all available options:
+ * npm run compare-performance -- -h
  *
  * The results will be saved in the folder `./benchmark/results`
  */
-const { spawn } = require('child_process');
-const myArgs = process.argv.slice(2);
-const prevVersion = myArgs.length === 1 ? '@' + myArgs[0] : '@latest';
+const { spawn, fork } = require('child_process');
 
-const benny = require('benny');
-const open = require('open');
-const current_fhirpath = require('../src/fhirpath');
-const current_r4_model = require('../fhir-context/r4');
-const minimumDataset = require('./resources/Minimum-Data-Set---version-3.0.R4.json');
-const _ = require('lodash');
-const currentVersion = 'current';
+// Insert performance test suites here:
+const availableTests = [
+  'intersect',
+  'member-invocation',
+  'union',
+  'exclude',
+  'subsetof',
+  'distinct',
+  'contains'
+];
 
+const options = require('commander');
+options.description('Compare performance between the latest published version and the current version.');
+options.option('-v, --prevVersion <version>', 'use a specific version instead of latest published version.', 'latest');
+options.option('-t, --tests <list>', `list of comma-separated tests (e.g. "${availableTests.join(',')}")`, availableTests.join(','));
+options.option('-c, --compileOnly', 'skip tests for evaluate().', 'latest');
+options.parse(process.argv);
 
-const child = spawn('npm', ['i','--prefix', './test/benchmark/prev-fhirpath', 'fhirpath' + prevVersion], {
+const npmInstallProcess = spawn('npm', ['i', '--prefix', './test/benchmark/prev-fhirpath', 'fhirpath@' + options.prevVersion], {
   stdio: 'inherit'
 });
 
-child.on('exit', code => {
-  if (code === 0) {
-    const previous_fhirpath = require('./benchmark/prev-fhirpath/node_modules/fhirpath');
-    const previous_r4_model = require('./benchmark/prev-fhirpath/node_modules/fhirpath/fhir-context/r4');
-    const previousVersion = require('./benchmark/prev-fhirpath/node_modules/fhirpath/package.json').version;
-    const bigItems = current_fhirpath.evaluate(minimumDataset,'repeat(item)', {},  current_r4_model);
-    const bigItemsCopy = _.cloneDeep(bigItems);
-    const smallItems = current_fhirpath.evaluate(minimumDataset,'repeat(item).repeat(code)', {},  current_r4_model)
-    const smallItemsCopy = _.cloneDeep(smallItems);
-    // Insert performance test suites here:
-    [
-      'intersect',
-      'member-invocation',
-      'union',
-      'exclude',
-      'subsetof',
-      'distinct',
-      'contains'
-    ].forEach(filename => {
-      const suites = require('./benchmark/'+ filename)({
-        benny,
-        open,
-        previous_fhirpath,
-        previous_r4_model,
-        current_fhirpath,
-        current_r4_model,
-        minimumDataset,
-        currentVersion,
-        previousVersion,
-        bigItems,
-        bigItemsCopy,
-        smallItems,
-        smallItemsCopy
-      });
-      suites.forEach(suite => {
-        const cases = suite.cases.reduce((arr, item) => {
-          arr.push(
-            benny.add(
-              `${item.name} [${previousVersion}]`,
-              item.testFunction.bind(
-                this,
-                previous_fhirpath,
-                previous_r4_model,
-                previous_fhirpath.compile(suite.expression, previous_r4_model)
-              )
-            )
-          );
-          arr.push(
-            benny.add(
-              `${item.name} [${currentVersion}]`,
-              item.testFunction.bind(
-                this,
-                current_fhirpath,
-                current_r4_model,
-                current_fhirpath.compile(suite.expression, current_r4_model)
-              )
-            )
-          );
-          return arr;
-        }, []);
+// Process for running benchmarks. We need a separate process to run the tests
+// to free the main process from synchronous code to listen for the SIGINT event.
+let benchmarkingProcess;
 
-        benny.suite(
-          suite.name,
-          ...cases,
-          benny.cycle(),
-          benny.complete(),
-          benny.configure({
-            minDisplayPrecision: 2
-          }),
-          benny.save({ file: suite.filename, folder: __dirname + '/results', version: currentVersion }),
-          benny.save({ file: suite.filename, folder: __dirname + '/results', format: 'chart.html' }),
-        ).then(() => {
-          open(__dirname + `/results/${suite.filename}.chart.html`);
-        });
-      });
-    })
+process.on('SIGINT', () => {
+  // Kill the benchmarking process
+  if (benchmarkingProcess) {
+    benchmarkingProcess.kill('SIGKILL');
+    // Displays the bash prompt on a new line.
+    console.log('');
+  }
+
+  // The value of the SIGINT signal code is 2 (See https://man7.org/linux/man-pages/man7/signal.7.html).
+  // If Node.js receives a fatal signal such as SIGKILL or SIGHUP, then its exit
+  // code will be 128 plus the value of the signal code:
+  // 128 + 2 = 130 (see https://nodejs.org/api/process.html#exit-codes)
+  process.exit(130);
+});
+
+npmInstallProcess.on('exit', code => {
+  if (code === 0) {
+    benchmarkingProcess = fork(__dirname + '/benchmark/runner.js', { stdio: 'inherit'});
+    // Pass options to the benchmarking process to run benchmarks
+    benchmarkingProcess.send(options);
   }
 });