+++ test/benchmark/distinct.js
@@ -4,6 +4,7 @@ const { createSuiteForExpression } = require('./common-benchmark-utils');
 module.exports = ({
                     bigItems,
                     smallItems,
+                    options
                   }) => {
   const expression = '%items.distinct()';
 
@@ -12,13 +13,15 @@ module.exports = ({
     filename: 'distinct',
     expression,
     bigItems,
-    smallItems
+    smallItems,
+    options
   }, {
     name: 'distinct() of a small collection with',
     filename: 'distinct-for-small-collections',
     expression,
     bigItems: bigItems.slice(-maxCollSizeForDeepEqual),
-    smallItems: smallItems.slice(-maxCollSizeForDeepEqual)
+    smallItems: smallItems.slice(-maxCollSizeForDeepEqual),
+    options
   }].map(createSuiteForExpression);
 
 }